refactor: move ExtendedCommit validation tests to blocksync package
Move the tests from types/extended_commit_poison_test.go to
blocksync/reactor_test.go so they live alongside the code they test.
